+++ projects/ngx-piano/src/lib/route/ngx-piano-route-meta-data.ts
@@ -1,0 +1,9 @@
+export interface NgxPianoRouteMetaData {
+  page?: string,
+  page_chapter1?: string,
+  page_chapter2?: string,
+  page_chapter3?: string,
+  page_chapter4?: string,
+  page_chapter5?: string,
+  page_chapter6?: string,
+}