+++ projects/ngx-piano/src/public-api.ts
@@ -4,5 +4,6 @@
 
 export * from './lib/ngx-piano.module';
 export * from './lib/event/ngx-piano-action-type';
+export * from './lib/route/ngx-piano-route-meta-data'
 export * from './lib/event/piano-track-click.directive';
 export * from './lib/piano-tracker';