+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add possibility to set common properties (key: value) for one or all events
 - Add possibility to let client project define which piano script it wants to use
+- Add possibility to add route meta-data to custom a page title of a route and others specifics meta-data 
 
 ### Changed